requirements.txt
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
langchain-groq
lanchain-community
plotly
-e .